[UPDATE] Auto cancel quotations
