--- autoremove_quotations/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import models
-from . import report
-from . import controllers
