--- autoremove_quotations/__manifest__.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-
-{
-    'name': 'Autoremove Quotations',
-    'summary': """ """,
-    'description': """
-    
-    """,
-    'author': 'Odoo',
-    'website': 'https://www.odoo.com',
-    'category': 'Training',
-    'version': '0.1',
-    'depends': ['sale','website','website_payment','l10n_mx_edi'],
-    'data': [
-        
-    ],
-    'demo': [
-        
-    ],
-    
-}
