--- autoremove_quotations/security/ir.model.access.csv
@@ -1,1 +1,0 @@
-id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
