+++ tvu_network/__manifest__.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+{
+    'name': 'Auto cancel',
+    'sumary': 'Auto cancel expired quotations',
+    'description': """
+        Auto cancel app to cancel expired quotations:
+    """,
+    'author': 'damu',
+    'category': 'Sales',
+    'version': '0.1',
+    'depends': ['sale'],
+    'license' : 'OPL-1',
+    'data': [
+    ],
+   'demo': [
+   ],
+    
+}
