+++ tvu_network/models/__init__.py
@@ -1,3 +1,2 @@
 # -*- coding: utf-8 -*-
-
 from . import sale_order
